backend/requirements.txt
# --- Git dependencies ---
BoSegmenter @git+https://github.com/OpenPecha/BoSegmenter.git@d31da9d9cc913ddb01e006446f66191078875d95
fast_antx @git+https://github.com/OpenPecha/fast-antx.git@4dad71a8340c3e400e59ab70027347698648f1b0
ai-text-outline @git+https://github.com/OpenPecha/ai-text-outline.git

# --- Core web stack ---
fastapi
uvicorn[standard]
mangum

# --- Async / IO ---
aiofiles
anyio
httpx

# --- Data / validation ---
pydantic[email]

# --- NLP / Tibetan ---
botok
pyewts

# --- Utilities ---
PyJWT[crypto]
python-dotenv
python-multipart
PyYAML
tqdm

# --- RDF / APIs ---
rdflib
requests

# --- Database ---
sqlalchemy
alembic
psycopg2-binary

# --- Cache ---
redis

# --- GenAI ---
google-genai